Add the start of a new GravityBlocks pattern. Only shows touched pixels in a random colour at the moment.
This pattern connected to a Node.js daemon which manages inputs from
multiple connected web clients. Currently, these web clients can touch
a square on an 8x8 grid, and these touches will be displayed on the
front face of the cube in a random colour.

Colour data is transmitted along with each touch, but it isn't used yet.

The pattern is currently single buffered, but conversion to double
buffered would be preferred.